feat(lumen): spec-faithful Temporal.Duration engine
Rework Temporal.Duration onto the spec's data model: components are
integral float64s (observable 𝔽(x) precision loss included), while all
exact reasoning happens in an InternalDuration record — a calendar date
part plus an i128 nanosecond time part bounded by 2^53 seconds.

- toString/toJSON format the seconds group exactly in i128 (huge
  components no longer wrap or misplace the decimal point), a blank
  duration prints PT0S, and toString now really rounds with
  smallestUnit/fractionalSecondDigits/roundingMode.
- round/total/compare gain full relativeTo support: plain dates,
  ZonedDateTime (DST-aware day spans via the zone's transition table),
  property bags (alphabetical single-read field order), and ISO strings
  (offset/annotation validation, instant range checks).
- Rounding follows the spec's nudge algorithms (NudgeToCalendarUnit /
  NudgeToZonedTime / NudgeToDayOrTime plus BubbleRelativeDuration) with
  exact rational boundary comparisons, including the tc39 #3141/#3148
  origin-anchored fixes for repeated wall-clock times at DST folds.
- Constructors honor a subclass new-target's prototype (observable
  GetPrototypeFromConstructor); negated()/round() normalize -0;
  toLocaleString delegates to Intl.DurationFormat.
- Number(hugeBigInt) is now correctly rounded (JsBigInt::to_f64 rounds
  to nearest/ties-to-even instead of truncating the top 64 bits).

built-ins/Temporal/Duration + intl402/Temporal/Duration: 561/561.

+++ crates/lumen/src/intl/durationformat.rs
@@ -302,16 +302,7 @@ fn read_duration(i: &mut Interp, v: &Value) -> Result<[f64; 10], Value> {
             i.temporal.get(&(std::rc::Rc::as_ptr(o) as usize))
         {
             return Ok([
-                d.years as f64,
-                d.months as f64,
-                d.weeks as f64,
-                d.days as f64,
-                d.hours as f64,
-                d.minutes as f64,
-                d.seconds as f64,
-                d.ms as f64,
-                d.us as f64,
-                d.ns as f64,
+                d.years, d.months, d.weeks, d.days, d.hours, d.minutes, d.seconds, d.ms, d.us, d.ns,
             ]);
         }
     }
